refactor: login register
