--- src/pages/Login/LoginForm/index.less
@@ -1,1 +1,0 @@
-/* @import '~antd/dist/antd.css'; */
