+++ src/pages/Login/index.less
@@ -1,13 +1,16 @@
-/* body{
-    background-color: aliceblue;
-} */
-.loginOuter {
-    /* height: 300px; */
+.login-form {
     width: 400px;
     margin: 150px auto;
-    padding: 60px;
+    padding: 40px;
     border: 1px solid rgba(0, 0, 0, 0.1);
     border-radius: 6px;
     background-color: white;
     box-shadow: 0 0 2px rgba(0, 0, 0, 0.2);
+    .form-title{
+        text-align: center;
+        margin-bottom: 30px;
+    }
+    .login-form-button{
+        margin-right: 20px;
+    }
 }
