--- src/pages/Register/RegisterForm/index.less
@@ -1,5 +1,0 @@
-/* @import '~antd/dist/antd.css'; */
-#vefifyCode {
-    display: flex;
-    gap: 10px;
-}
