+++ src/pages/Register/index.less
@@ -1,13 +1,5 @@
-/* body{
-    background-color: aliceblue;
-} */
-.registerOuter {
-    /* height: fit-content; */
-    width: 400px;
-    margin: 150px auto;
-    padding: 60px;
-    border: 1px solid rgba(0, 0, 0, 0.1);
-    border-radius: 6px;
-    background-color: white;
-    box-shadow: 0 0 2px rgba(0, 0, 0, 0.2);
+#vefifyCode {
+    display: flex;
+    gap: 10px;
 }
+
